Rename incorrectly cased files on updates or when fixing database (#32006)
* Case sensitive rename

* Add comment

* Add si_LK.gif

* Rename files

* Correct class name

* Move filename casing into the delete files method

* Fix for OSX

* Fix comment
