+++ libraries/src/Filesystem/Support/StringController.php
